pyproject.toml
[project]
name = "rag-chat-bot"
version = "0.1.0"
description = "Diet food RAG chatbot with memory bank"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "openai>=1.60.0",
    "supabase>=2.11.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "tiktoken>=0.8.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "pymupdf>=1.24.0",
]

[project.scripts]
chat = "cli:main"
